+++ zcash_client_sqlite/CHANGELOG.md
@@ -10,6 +10,13 @@ workspace.
 
 ## [Unreleased]
 
+### Added
+- The following columns have been added to the exposed `v_tx_outputs` view:
+  - `transaction_id`
+  - `tx_mined_height`
+  - `tx_trust_status`
+  - `recipient_key_scope`
+
 ### Changed
 - Migrated to `orchard 0.12`, `sapling-crypto 0.6`.